refactor: rename project, update bundle ID, README and pre-commit logic
- Changed app name to "Gestione Donazione AVIS"
- Updated bundle identifier to com.ddd.avis_donation_management_app
- Renamed main application widget to match new app name
- Created English version of the README (README.en.md)
- Replaced "Ingresso"/"Entry" with "Accoglienza"/"Reception" in both README files and in the donation page
- Updated "Technologies Used" section to reflect actual dependencies
- Added language switch links between Italian and English READMEs
- Aligned terminology and structure across both README versions
- Refined project folder structure: added components/, adjusted helpers/
- Updated pre-commit hook to process badge updates on multiple README files
