+++ lib/components/avis_scaffold.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:avis_donor_app/helpers/app_info_controller.dart';
-import 'package:avis_donor_app/helpers/connection_status_controller.dart';
-import 'package:avis_donor_app/helpers/operator_session_controller.dart';
-import 'package:avis_donor_app/helpers/avis_drawer.dart';
-import 'package:avis_donor_app/helpers/avis_bottom_navigation_bar.dart';
+import 'package:avis_donation_management/helpers/app_info_controller.dart';
+import 'package:avis_donation_management/helpers/connection_status_controller.dart';
+import 'package:avis_donation_management/helpers/operator_session_controller.dart';
+import 'package:avis_donation_management/components/avis_drawer.dart';
+import 'package:avis_donation_management/components/avis_bottom_navigation_bar.dart';
 
 /// Reusable Scaffold with AVIS AppBar, Drawer, and optional BottomNavigationBar
 class AvisScaffold extends StatelessWidget {
