+++ pubspec.yaml
@@ -1,4 +1,4 @@
-name: avis_donor_app
+name: avis_donation_management
 description: "Applicazione per la gestione delle donazioni."
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
